magnum-v2/requirements.txt
Flask==3.0.3
openstacksdk==4.10.0
python-dotenv==1.0.1
requests==2.32.3
Flask-CORS==4.0.1
python-dateutil==2.9.0
